+++ src/Inputtape.cpp
@@ -7,12 +7,12 @@
   InputTape::InputTape(string fileName){
     index = 0;
     fstream myfile(fileName.c_str());
-    int dummy;
+    float dummy;
     while(myfile >> dummy)
       inputtape.push_back(dummy);
   } 
 
-  int InputTape::read(void){
+  float InputTape::read(void){
     int temp = index;
     index++;
     return inputtape[temp];
