+++ src/OutputUnit.h
@@ -21,7 +21,7 @@ public:
   *
   * Este metodo escribe un valor en la cinta de salida
   */
-  void write(int data);
+  void write(float data);
 
 /** \brief toFile
   *
